+++ Configuration/ProcessModifiers/python/deepTauSonicTriton_cff.py
@@ -1,0 +1,3 @@
+import FWCore.ParameterSet.Config as cms
+
+deepTauSonicTriton = cms.Modifier()
