+++ RecoTauTag/RecoTau/BuildFile.xml
@@ -25,6 +25,7 @@
 <use name="CommonTools/BaseParticlePropagator"/>
 <use name="RecoLocalCalo/HGCalRecAlgos"/>
 <use name="roottmva"/>
+<use name="PhysicsTools/TensorFlow"/>
 <export>
   <lib name="1"/>
 </export>
